這次又更新了很多

+++ app/Domains/Auth/Http/Controllers/Backend/Role/RoleController.php
@@ -2,6 +2,9 @@
 
 namespace App\Domains\Auth\Http\Controllers\Backend\Role;
 
+use App\Domains\Auth\Events\Role\RoleCreated;
+use App\Domains\Auth\Events\Role\RoleDeleted;
+use App\Domains\Auth\Events\Role\RoleUpdated;
 use App\Domains\Auth\Http\Requests\Backend\Role\DeleteRoleRequest;
 use App\Domains\Auth\Http\Requests\Backend\Role\EditRoleRequest;
 use App\Domains\Auth\Http\Requests\Backend\Role\StoreRoleRequest;
@@ -9,6 +12,7 @@ use App\Domains\Auth\Http\Requests\Backend\Role\UpdateRoleRequest;
 use App\Domains\Auth\Models\Role;
 use App\Domains\Auth\Repositories\PermissionRepository;
 use App\Domains\Auth\Repositories\RoleRepository;
+use App\Exceptions\GeneralException;
 use App\Http\Controllers\Controller;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Http\RedirectResponse;
@@ -43,7 +47,9 @@ class RoleController extends Controller
 
     public function store(StoreRoleRequest $request): Redirector|RedirectResponse
     {
-        $this->roleRepository->createOrUpdateFromArray($request->validated());
+        $role = $this->roleRepository->createOrUpdateFromArray($request->validated());
+
+        event(new RoleCreated($role));
 
         return redirect()
             ->route('admin.auth.role.index')
@@ -53,15 +59,17 @@ class RoleController extends Controller
     public function edit(EditRoleRequest $request, Role $role): Factory|View
     {
         return view('backend.auth.role.edit')
-            ->with('categories', $this->permissionService->getCategorizedPermissions())
-            ->with('general', $this->permissionService->getUncategorizedPermissions())
+            ->with('categories', $this->permissionRepository->getCategorizedPermissions())
+            ->with('general', $this->permissionRepository->getUncategorizedPermissions())
             ->with('role', $role)
             ->with('usedPermissions', $role->permissions->modelKeys());
     }
 
     public function update(UpdateRoleRequest $request, Role $role): Redirector|RedirectResponse
     {
-        $this->roleRepository->updateByPrimary($role->id, $request->validated());
+        $this->roleRepository->updateByPrimary($role->id, $request->validated(), false);
+
+        event(new RoleUpdated($role));
 
         return redirect()
             ->route('admin.auth.role.index')
@@ -70,7 +78,13 @@ class RoleController extends Controller
 
     public function destroy(DeleteRoleRequest $request, Role $role): Redirector|RedirectResponse
     {
+        if ($role->users()->count()) {
+            throw new GeneralException(__('You can not delete a role with associated users.'));
+        }
+
         $this->roleRepository->deleteByPrimary($role->id);
+
+        event(new RoleDeleted($role));
 
         return redirect()
             ->route('admin.auth.role.index')
